parse-json: extract articles from JSON files.
The articles are put in subdirectories of the articles/ directory,
using the SHA1 hash of the raw text to determine the filename. The
first 3 characters of the hash (in hex format) are the subdirectory
and the rest are the basename. Like so:

    xxx/yyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyy

There are around 4 million articles and 4096 directories, hence about
1000 articles in each directory. Each directory is of course a random
sample.

The article files are encoded in utf-8. They start with a header
containing metadata (article title, page url, issue title, issue code,
article type, sha1). After the header comes a single blank line, then
the article content contained within Start-of-text and end-of-text
markers (␂, ␃). Uncommon and unwanted characters (likely the product
of OCR noise) have been replaced with the ° symbol.