preprocessing: tougher normalisation of whitespace
It turns out that some of the example documents had terrible formatting,
and there is not much else that can be done. Also, the TEI <p> tags
presumably are supposed to be completely normalised in their formatting,
like html.

TEI <note>s are in-line but out of stream comments that should not be
included.