Add options for periodic image dumps of RNN weights
The syntax is rather obscure and defined in the recur repository.

     ./train-net-multi --periodic-pgm-dump "ihw how ihd" \
                       --periodic-pgm-period=10000

will dump the input->hidden weights (including recurrent hidden->hidden
weights) because "ihw", the hidden->output weights ("how"), and the
input->hidden delta ("ihd").

Other options for the third character are "a" (aux), "m" (momentum),
and sometimes "t" (tmp). It depends on the learning method.