fix missing soft-hyphens in parse-json.
Where the text was ending a line with a hyphen, the JSON mistakenly
had a double space. Thus an article like this:

    CARGO OF GERMAN RIFLES.
    COPENHAGEN, April 1
    There  is  much  speculation regard-
    ing the sudden departure of the Nor-
    wegian  steamer  Fanny  from  Longe-
    land  with  300  tons   of   rifles,
    [...]

was being translated into this:

    CARGO OF GERMAN RIFLES.
    COPENHAGEN, April 1
    There is much speculation regard  ing the sudden departure of the
    Nor  wegian steamer Fanny from Longe  land with 300 tons of rifles,

So here we replace all double spaces with *no* spaces, joining the words.
It seems there aren't any deliberate double spaces, but there are a few
words that would normally be hyphenated when not at a line ending which
lose their hyphens though this process (e.g., "beach-bathing",
"out-buildings", "twenty-pound").