Script to filter articles according to patterns
This produces versions of results files filtered by file specific
regular expressions. '!' and '-' are special patterns both meaning
"don't filter on this one". This:

./filter-results articles results/scores-00x.txt '!' \
                          results/papers-00x.txt '!' \
                          results/entropy-00x.txt '-' \
                          results/types-00x.txt  '^Art' \
                          results/trigram-tfidf-3-6-3000-00x.txt '!' \
                          results/years-00x.txt '!'

means produce a set of files with the suffix '-articles' (from the first
argument) which contains every line of the corresponding '.txt' file IF
the line in 'results/types-00x.txt' starts with 'Art'.